Updated Rhino and HTMLUnit Core
git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@33991 ebf7c1c2-ba36-0410-9fe8-c592906822b4

Former-commit-id: 45863817288bcbfe7112bcc5c835e87938cae48f

+++ src/org/jdownloader/phantomjs/installation/InstallThread.java
@@ -221,7 +221,9 @@ public class InstallThread extends Thread {
                 StatsManager.I().track("installing/result/" + new PhantomJS().isAvailable(), CollectionName.PJS);
                 installing = false;
                 file.delete();
-                Files.deleteRecursiv(extractTo);
+                if (extractTo.exists()) {
+                    Files.deleteRecursiv(extractTo);
+                }
             }
         } catch (Throwable e) {
             LogController.CL(false).log(e);
